requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
openai==0.28.0  # Using older version for compatibility with the code
python-dotenv==1.0.0
pydantic==2.4.2  # Specific version to avoid conflicts
# pydantic-settings removed as we're not using it anymore
pytest==7.4.3
httpx==0.25.1
python-jose[cryptography]==3.3.0

# AI model dependencies
transformers>=4.34.0  # Required for Hugging Face models (4.34+ needed for TinyLlama)
torch>=2.0.0  # Required for Hugging Face models
sentencepiece>=0.1.99  # Required for some Hugging Face models
accelerate>=0.20.3  # For optimized model loading
einops>=0.6.1  # Required for TinyLlama
bitsandbytes>=0.41.0  # For quantization support (optional but recommended)
